Action sub results.
Add the field sub_results to the action struct. It is for side effects that
happens on top of the action result.

Add "Unit Enter Hut" and "Unit Frighten Hut" as the first action sub
results. Move the parts of the corresponding action results that can be
moved to the action sub results.

See osdn ##41610