Introduce req[_vec]_is_impossible_to_fulfill().
Introduce the new functions req_is_impossible_to_fulfill() and
req_vec_is_impossible_to_fulfill(). They check if a requirement or
requirement vector is known to be impossible to fulfill in the current
ruleset.

Introduce the new function universal_never_there() to check if a universal
is known to never be there.

The only currently supported requirement types are action, unit class flag
and unit type flag.

See osdn ##41613